workspace/02_工程/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
pillow>=10.0.0
python-multipart>=0.0.9
